style: apply cargo fmt and fix clippy lints
Normalize formatting across the crate with cargo fmt and resolve all
clippy findings under -D warnings ahead of enabling lint enforcement in CI.

Genuine fixes:
- remove a duplicated #![deny(unsafe_code)] in main.rs
- repeat.rs: index loop -> iterator
- track.rs: manual chrom[3..] -> strip_prefix("chr")
- gene_annotations.rs: fix overindented doc list
- tests: drop an orphaned doc-comment fragment; rewrite
  str[a..b].as_bytes() -> &str.as_bytes()[a..b]

Deliberate allows (domain-driven thresholds, not worth signature churn),
added crate-level in main.rs with rationale:
- too_many_arguments (builder/tally functions)
- type_complexity (intermediate tuple collections)
- large_enum_variant (clap Command subcommand enum)

+++ src/fragments/mod.rs
@@ -5,7 +5,7 @@ use tracing::info;
 use crate::bam::{gc_frac, open_bam, read_group_sample_id, RefCache};
 use crate::cli::FragmentsArgs;
 use crate::record::FragmentRecord;
-use crate::region::{RegionInput, parse_region_input};
+use crate::region::{parse_region_input, RegionInput};
 use crate::writer::parquet_fragments::FragmentsWriter;
 
 pub fn collect_fragments(args: &FragmentsArgs, writer: &mut FragmentsWriter) -> Result<()> {
@@ -46,7 +46,9 @@ pub fn collect_fragments(args: &FragmentsArgs, writer: &mut FragmentsWriter) -> 
 
     for fetch_region in &fetch_regions {
         match fetch_region.as_deref() {
-            Some(r) => bam.fetch(r).with_context(|| format!("failed to fetch region '{r}'"))?,
+            Some(r) => bam
+                .fetch(r)
+                .with_context(|| format!("failed to fetch region '{r}'"))?,
             None => bam.fetch(".").context("failed to fetch all reads")?,
         }
 
@@ -65,7 +67,7 @@ pub fn collect_fragments(args: &FragmentsArgs, writer: &mut FragmentsWriter) -> 
                 || !record.is_proper_pair()
                 || record.is_secondary()
                 || record.is_supplementary()
-                || (record.mapq() as u8) < args.min_map_qual
+                || record.mapq() < args.min_map_qual
             {
                 continue;
             }
